Changing solr loading commands to fail fast if solr exceptions are thrown
Previously these were caught and printed out on stderr. This is not so good as
they are buried in the masses of log messages meaning problems leak into production
potentially

Changed behavior so that exceptions are not caught. This will cause the current command
to die, and to exit with a non-zero error code. Fixes #223

Note that in two places we still wrap with a try{}, in the experimental model loading.
However, the behavior has been changed here to explicitly exit with non-zero if
any exceptions are caught